+++ src/challenger.rs
@@ -207,17 +207,11 @@ mod tests {
 
     use std::sync::mpsc::{channel, Receiver, Sender};
 
-    use bitcoin_hashes::Hash;
-
     use crate::clientchain::MockClientChain;
     use crate::error::Error;
     use crate::service::MockService;
     use crate::storage::MockStorage;
-
-    /// Generate dummy hash for tests
-    fn gen_dummy_hash(i: u8) -> sha256d::Hash {
-        sha256d::Hash::from_slice(&[i as u8; 32]).unwrap()
-    }
+    use crate::testing_utils::tests::gen_dummy_hash;
 
     #[test]
     fn verify_challenge_test() {
