+++ src/lib.rs
@@ -44,3 +44,4 @@ pub mod ocean;
 pub mod request;
 pub mod service;
 pub mod storage;
+pub mod testing_utils;
